Enable warnings and C++17 compiler settings
This change is a mix of warning fixes from Clang, clang-tidy fixes, and three nolints.

The nolints were added because sqrt() is being used by a noexcept function to take the
square root of a positive constant. Because sqrt() only throws when receiving a negative
input, the tool is generating a false positive.